+++ src/slg/film/filters/gaussian.cpp
@@ -40,21 +40,21 @@ Properties GaussianFilter::ToProperties(const Properties &cfg) {
 }
 
 Filter *GaussianFilter::FromProperties(const Properties &cfg) {
-	const float defaultFilterWidth = cfg.Get(GetDefaultProps().Get("film.filter.width")).Get<float>();
-	const float filterXWidth = cfg.Get(Property("film.filter.xwidth")(defaultFilterWidth)).Get<float>();
-	const float filterYWidth = cfg.Get(Property("film.filter.ywidth")(defaultFilterWidth)).Get<float>();
+	const float defaultFilterWidth = cfg.Get(GetDefaultProps().Get("film.filter.width")).Get<double>();
+	const float filterXWidth = cfg.Get(Property("film.filter.xwidth")(defaultFilterWidth)).Get<double>();
+	const float filterYWidth = cfg.Get(Property("film.filter.ywidth")(defaultFilterWidth)).Get<double>();
 
-	const float alpha = cfg.Get(GetDefaultProps().Get("film.filter.gaussian.alpha")).Get<float>();
+	const float alpha = cfg.Get(GetDefaultProps().Get("film.filter.gaussian.alpha")).Get<double>();
 
 	return new GaussianFilter(filterXWidth, filterYWidth, alpha);
 }
 
 slg::ocl::Filter *GaussianFilter::FromPropertiesOCL(const Properties &cfg) {
-	const float defaultFilterWidth = cfg.Get(GetDefaultProps().Get("film.filter.width")).Get<float>();
-	const float filterXWidth = cfg.Get(Property("film.filter.xwidth")(defaultFilterWidth)).Get<float>();
-	const float filterYWidth = cfg.Get(Property("film.filter.ywidth")(defaultFilterWidth)).Get<float>();
+	const float defaultFilterWidth = cfg.Get(GetDefaultProps().Get("film.filter.width")).Get<double>();
+	const float filterXWidth = cfg.Get(Property("film.filter.xwidth")(defaultFilterWidth)).Get<double>();
+	const float filterYWidth = cfg.Get(Property("film.filter.ywidth")(defaultFilterWidth)).Get<double>();
 
-//	const float alpha = cfg.Get(GetDefaultProps().Get("film.filter.gaussian.alpha")).Get<float>();
+//	const float alpha = cfg.Get(GetDefaultProps().Get("film.filter.gaussian.alpha")).Get<double>();
 
 	slg::ocl::Filter *oclFilter = new slg::ocl::Filter();
