+++ src/slg/film/filters/mitchell.cpp
@@ -42,23 +42,23 @@ Properties MitchellFilter::ToProperties(const Properties &cfg) {
 }
 
 Filter *MitchellFilter::FromProperties(const Properties &cfg) {
-	const float defaultFilterWidth = cfg.Get(GetDefaultProps().Get("film.filter.width")).Get<float>();
-	const float filterXWidth = cfg.Get(Property("film.filter.xwidth")(defaultFilterWidth)).Get<float>();
-	const float filterYWidth = cfg.Get(Property("film.filter.ywidth")(defaultFilterWidth)).Get<float>();
+	const float defaultFilterWidth = cfg.Get(GetDefaultProps().Get("film.filter.width")).Get<double>();
+	const float filterXWidth = cfg.Get(Property("film.filter.xwidth")(defaultFilterWidth)).Get<double>();
+	const float filterYWidth = cfg.Get(Property("film.filter.ywidth")(defaultFilterWidth)).Get<double>();
 
-	const float b = cfg.Get(GetDefaultProps().Get("film.filter.mitchell.b")).Get<float>();
-	const float c = cfg.Get(GetDefaultProps().Get("film.filter.mitchell.c")).Get<float>();
+	const float b = cfg.Get(GetDefaultProps().Get("film.filter.mitchell.b")).Get<double>();
+	const float c = cfg.Get(GetDefaultProps().Get("film.filter.mitchell.c")).Get<double>();
 
 	return new MitchellFilter(filterXWidth, filterYWidth, b, c);
 }
 
 slg::ocl::Filter *MitchellFilter::FromPropertiesOCL(const Properties &cfg) {
-	const float defaultFilterWidth = cfg.Get(GetDefaultProps().Get("film.filter.width")).Get<float>();
-	const float filterXWidth = cfg.Get(Property("film.filter.xwidth")(defaultFilterWidth)).Get<float>();
-	const float filterYWidth = cfg.Get(Property("film.filter.ywidth")(defaultFilterWidth)).Get<float>();
+	const float defaultFilterWidth = cfg.Get(GetDefaultProps().Get("film.filter.width")).Get<double>();
+	const float filterXWidth = cfg.Get(Property("film.filter.xwidth")(defaultFilterWidth)).Get<double>();
+	const float filterYWidth = cfg.Get(Property("film.filter.ywidth")(defaultFilterWidth)).Get<double>();
 
-//	const float b = cfg.Get(GetDefaultProps().Get("film.filter.mitchell.b")).Get<float>();
-//	const float c = cfg.Get(GetDefaultProps().Get("film.filter.mitchell.c")).Get<float>();
+//	const float b = cfg.Get(GetDefaultProps().Get("film.filter.mitchell.b")).Get<double>();
+//	const float c = cfg.Get(GetDefaultProps().Get("film.filter.mitchell.c")).Get<double>();
 
 	slg::ocl::Filter *oclFilter = new slg::ocl::Filter();
