+++ src/slg/film/filters/mitchellss.cpp
@@ -42,23 +42,23 @@ Properties MitchellSSFilter::ToProperties(const Properties &cfg) {
 }
 
 Filter *MitchellSSFilter::FromProperties(const Properties &cfg) {
-	const float defaultFilterWidth = cfg.Get(GetDefaultProps().Get("film.filter.width")).Get<float>();
-	const float filterXWidth = cfg.Get(Property("film.filter.xwidth")(defaultFilterWidth)).Get<float>();
-	const float filterYWidth = cfg.Get(Property("film.filter.ywidth")(defaultFilterWidth)).Get<float>();
+	const float defaultFilterWidth = cfg.Get(GetDefaultProps().Get("film.filter.width")).Get<double>();
+	const float filterXWidth = cfg.Get(Property("film.filter.xwidth")(defaultFilterWidth)).Get<double>();
+	const float filterYWidth = cfg.Get(Property("film.filter.ywidth")(defaultFilterWidth)).Get<double>();
 
-	const float b = cfg.Get(GetDefaultProps().Get("film.filter.mitchellss.b")).Get<float>();
-	const float c = cfg.Get(GetDefaultProps().Get("film.filter.mitchellss.c")).Get<float>();
+	const float b = cfg.Get(GetDefaultProps().Get("film.filter.mitchellss.b")).Get<double>();
+	const float c = cfg.Get(GetDefaultProps().Get("film.filter.mitchellss.c")).Get<double>();
 
 	return new MitchellSSFilter(filterXWidth, filterYWidth, b, c);
 }
 
 slg::ocl::Filter *MitchellSSFilter::FromPropertiesOCL(const Properties &cfg) {
-	const float defaultFilterWidth = cfg.Get(GetDefaultProps().Get("film.filter.width")).Get<float>();
-	const float filterXWidth = cfg.Get(Property("film.filter.xwidth")(defaultFilterWidth)).Get<float>();
-	const float filterYWidth = cfg.Get(Property("film.filter.ywidth")(defaultFilterWidth)).Get<float>();
+	const float defaultFilterWidth = cfg.Get(GetDefaultProps().Get("film.filter.width")).Get<double>();
+	const float filterXWidth = cfg.Get(Property("film.filter.xwidth")(defaultFilterWidth)).Get<double>();
+	const float filterYWidth = cfg.Get(Property("film.filter.ywidth")(defaultFilterWidth)).Get<double>();
 
-//	const float b = cfg.Get(GetDefaultProps().Get("film.filter.mitchellss.b")).Get<float>();
-//	const float c = cfg.Get(GetDefaultProps().Get("film.filter.mitchellss.c")).Get<float>();
+//	const float b = cfg.Get(GetDefaultProps().Get("film.filter.mitchellss.b")).Get<double>();
+//	const float c = cfg.Get(GetDefaultProps().Get("film.filter.mitchellss.c")).Get<double>();
 
 	slg::ocl::Filter *oclFilter = new slg::ocl::Filter();
